backend-python/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pymongo==4.6.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==1.10.13
aiofiles==23.2.1
Pillow==10.1.0
python-magic==0.4.27
httpx==0.25.2
slowapi==0.1.9
email-validator==2.2.0 